pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "security-mcp"
version = "1.0.0"
description = "security-mcp — Security Testing MCP Server"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "aiosqlite>=0.20,<1",
    "beautifulsoup4>=4.12,<5",
    "cryptography>=43,<44",
    "dnspython>=2.6,<3",
    "httpx>=0.27,<1",
    "litellm>=1.50,<2",
    "playwright>=1.48",
    "pydantic>=2.5,<3",
    "pyyaml>=6.0,<7",
    "rank-bm25>=0.2,<1",
    "zstandard>=0.23,<1",
]

[project.optional-dependencies]
mcp = ["mcp[cli]>=1.26.0"]
vault = ["hvac>=2.4"]
protocols = [
    "asyncssh>=2.14",
    "smbprotocol>=1.12",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
]
all = [
    "security-mcp[mcp,vault,protocols,dev]",
]

[project.scripts]
security-mcp = "security_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/security_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["security_mcp"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "benchmark: marks benchmark tests",
    "integration: marks integration tests",
]
